// extensions/color-picker/src/lib/types.ts
export type Color = {
  alpha: number;
  red: number; // between 0 and 1
  green: number; // between 0 and 1
  blue: number; // between 0 and 1
  colorSpace: string;
};

export type DeprecatedColor = {
  alpha: number;
  red: number; // between 0 and 255
  green: number; // between 0 and 255
  blue: number; // between 0 and 255
};

export type HistoryColor = Color | DeprecatedColor | string;

export type HistoryItem = {
  date: string;
  color: HistoryColor;
  title?: string;
  isFavorite?: boolean;
};

export type SortType = "platform" | "proximity";

export type ColorFormatType =
  | "hex"
  | "hex-lower-case"
  | "hex-no-prefix"
  | "rgb"
  | "rgb-percentage"
  | "rgba"
  | "rgba-percentage"
  | "hsla"
  | "hsva"
  | "oklch"
  | "lch"
  | "p3";

export type UseColorsSelectionObject<T = HistoryItem | string> = {
  actions: {
    toggleSelection: (item: T) => void;
    selectAll: () => void;
    clearSelection: () => void;
  };
  selected: {
    selectedItems: T[];
    anySelected: boolean;
    allSelected: boolean;
    countSelected: number;
  };
  helpers: {
    getIsItemSelected: (item: T) => boolean;
  };
};

export type CopyColorsFormat = "json" | "css-classes" | "css-variables";

export type SelectMode = "single" | "multi";

export interface ExtensionPreferences {
  colorFormat?: ColorFormatType;
}

export interface PickColorPreferences {
  showColorName?: boolean;
}

export interface OrganizeColorsPreferences {
  primaryAction?: "copy" | "paste";
}

export interface ColorNamesPreferences {
  colorNamesPerGroup?: string;
}